float working, added automatic map size reading

+++ functions/planner.c
@@ -6,8 +6,8 @@
 #include "common.h"
 
 
-void planners_2D(double* matriz, int* size_map, double* objective_points, int size_objective[2], 
-                double* start_points, int size_start[2], int planner_type, int escalado_vectores) {
+void planners_2D(float* matriz, int* size_map, float* objective_points, int size_objective[2], 
+                float* start_points, int size_start[2], int planner_type, int escalado_vectores) {
 
     // Choose type of planner
     switch(planner_type){        
@@ -40,7 +40,7 @@ void planners_2D(double* matriz, int* size_map, double* objective_points, int si
                 if (y < y_min) y_min = y;
             }
             
-            double distance = euclidean_distance(x_max, y_max, x_min, y_min);
+            float distance = euclidean_distance(x_max, y_max, x_min, y_min);
             // Calculate margin as 10% of the distance
             int margin = (int)(0.1 * distance);
             if (margin < 2) margin = 2;  // Minimum margin of 2 cells
@@ -77,27 +77,27 @@ void planners_2D(double* matriz, int* size_map, double* objective_points, int si
             int y_min = size_map[1];
 
             for (int i = 0; i<num_start; i++){
-                double start_x = start_points[i*2];
-                double start_y = start_points[i*2 + 1];
+                float start_x = start_points[i*2];
+                float start_y = start_points[i*2 + 1];
 
                 for (int j = 0; j<num_end; j++){
-                    double end_x = objective_points[j*2];
-                    double end_y = objective_points[j*2 + 1];
+                    float end_x = objective_points[j*2];
+                    float end_y = objective_points[j*2 + 1];
 
                     //Calculate vector
-                    double vector_x = end_x - start_x;
-                    double vector_y = end_y - start_y;
+                    float vector_x = end_x - start_x;
+                    float vector_y = end_y - start_y;
 
                     //Normalize
-                    double norm = sqrt(vector_x*vector_x + vector_y*vector_y);
+                    float norm = sqrt(vector_x*vector_x + vector_y*vector_y);
                     vector_x = (vector_x / norm) * escalado_vectores; 
                     vector_y = (vector_y / norm) * escalado_vectores;
 
-                    double normal_x = -vector_y;
-                    double normal_y = vector_x;
+                    float normal_x = -vector_y;
+                    float normal_y = vector_x;
 
                     // Check points for start position
-                    double points[8][2] = {
+                    float points[8][2] = {
                         {start_x + vector_x, start_y + vector_y},   // Forward
                         {start_x - vector_x, start_y - vector_y},   // Backward
                         {start_x + normal_x, start_y + normal_y},   // Right
@@ -156,8 +156,8 @@ void planners_2D(double* matriz, int* size_map, double* objective_points, int si
     }
 }
 
-void planners_3D(double* matriz, int* size_map, double* objective_points, int size_objective[2], 
-    double* start_points, int size_start[2], int planner_type, int escalado_vectores){
+void planners_3D(float* matriz, int* size_map, float* objective_points, int size_objective[2], 
+    float* start_points, int size_start[2], int planner_type, int escalado_vectores){
         
     switch(planner_type){        
         case 0:
@@ -174,36 +174,36 @@ void planners_3D(double* matriz, int* size_map, double* objective_points, int si
             printf("\nCalculating 3D corridor bounds...\n");
 
             for (int i = 0; i < num_start; i++) {
-                double start_x = start_points[i*3];
-                double start_y = start_points[i*3 + 1];
-                double start_z = start_points[i*3 + 2];
+                float start_x = start_points[i*3];
+                float start_y = start_points[i*3 + 1];
+                float start_z = start_points[i*3 + 2];
 
                 for (int j = 0; j < num_end; j++) {
-                    double end_x = objective_points[j*3];
-                    double end_y = objective_points[j*3 + 1];
-                    double end_z = objective_points[j*3 + 2];
+                    float end_x = objective_points[j*3];
+                    float end_y = objective_points[j*3 + 1];
+                    float end_z = objective_points[j*3 + 2];
 
                     // Calculate directing vector and normalize+scale in one step
-                    double vector_x = end_x - start_x;
-                    double vector_y = end_y - start_y;
-                    double vector_z = end_z - start_z;
-                    double norm = sqrt(vector_x*vector_x + vector_y*vector_y + vector_z*vector_z);
+                    float vector_x = end_x - start_x;
+                    float vector_y = end_y - start_y;
+                    float vector_z = end_z - start_z;
+                    float norm = sqrt(vector_x*vector_x + vector_y*vector_y + vector_z*vector_z);
                     vector_x = (vector_x / norm) * escalado_vectores;
                     vector_y = (vector_y / norm) * escalado_vectores;
                     vector_z = (vector_z / norm) * escalado_vectores;
 
                     // First perpendicular vector (already inherits scale from vector)
-                    double perp1_x = -vector_y;
-                    double perp1_y = vector_x;
-                    double perp1_z = 0;
+                    float perp1_x = -vector_y;
+                    float perp1_y = vector_x;
+                    float perp1_z = 0;
 
                     // Second perpendicular vector using cross product
-                    double perp2_x = vector_y*perp1_z - vector_z*perp1_y;
-                    double perp2_y = vector_z*perp1_x - vector_x*perp1_z;
-                    double perp2_z = vector_x*perp1_y - vector_y*perp1_x;
+                    float perp2_x = vector_y*perp1_z - vector_z*perp1_y;
+                    float perp2_y = vector_z*perp1_x - vector_x*perp1_z;
+                    float perp2_z = vector_x*perp1_y - vector_y*perp1_x;
 
                     // Generate points array
-                    double points[12][3] = {
+                    float points[12][3] = {
                         // Start point expansions
                         {start_x + vector_x, start_y + vector_y, start_z + vector_z},
                         {start_x - vector_x, start_y - vector_y, start_z - vector_z},
